mm, treewide: make addr parameter of free_pages() void *
Most callers of free_pages() and free_page() cast a pointer type to
unsigned long.

Make addr parameter of free_pages() and free_page() 'void *' and update
call sites with the following semantic patch and a bit of manual formatting
tweaking.

    @@
    identifier addr;
    expression order;
    @@
    (
    - free_pages(addr, order)
    + free_pages((void *)addr, order)
    |
    - free_page(addr)
    + free_page((void *)addr)
    )

    @@
    type T != void *;
    expression addr, order;
    @@
    (
    - free_pages((T)addr, order)
    + free_pages(addr, order)
    |
    - free_page((T)addr)
    + free_page(addr)
    )

+++ arch/powerpc/kvm/book3s_xive.c
@@ -1845,8 +1845,7 @@ void kvmppc_xive_cleanup_vcpu(struct kvm_vcpu *vcpu)
 
 		xive_native_disable_queue(xc->vp_id, q, i);
 		if (q->qpage) {
-			free_pages((unsigned long)q->qpage,
-				   xive->q_page_order);
+			free_pages(q->qpage, xive->q_page_order);
 			q->qpage = NULL;
 		}
 	}
