+++ drivers/char/xillybus/xillyusb.c
@@ -416,7 +416,7 @@ retry:
 
 memfail:
 	for (i--; i >= 0; i--)
-		free_pages((unsigned long)fifo->mem[i], buf_order);
+		free_pages(fifo->mem[i], buf_order);
 
 	kfree(fifo->mem);
 	fifo->mem = NULL;
@@ -437,7 +437,7 @@ static void fifo_mem_release(struct xillyfifo *fifo)
 		return;
 
 	for (i = 0; i < fifo->bufnum; i++)
-		free_pages((unsigned long)fifo->mem[i], fifo->buf_order);
+		free_pages(fifo->mem[i], fifo->buf_order);
 
 	kfree(fifo->mem);
 }
@@ -476,7 +476,7 @@ static void endpoint_dealloc(struct xillyusb_endpoint *ep)
 		struct xillybuffer *xb =
 			list_entry(this, struct xillybuffer, entry);
 
-		free_pages((unsigned long)xb->buf, ep->order);
+		free_pages(xb->buf, ep->order);
 		kfree(xb);
 	}
