+++ drivers/gpu/drm/msm/msm_iommu.c
@@ -423,7 +423,7 @@ msm_iommu_pagetable_free_pt(void *cookie, void *data, size_t size)
 	struct msm_iommu_pagetable *pagetable = cookie;
 
 	if (unlikely(pagetable->root_page_table == data)) {
-		free_pages((unsigned long)data, get_order(size));
+		free_pages(data, get_order(size));
 		pagetable->root_page_table = NULL;
 		return;
 	}
