+++ drivers/input/misc/xen-kbdfront.c
@@ -415,7 +415,7 @@ static void xenkbd_remove(struct xenbus_device *dev)
 		input_unregister_device(info->ptr);
 	if (info->mtouch)
 		input_unregister_device(info->mtouch);
-	free_page((unsigned long)info->page);
+	free_page(info->page);
 	kfree(info);
 }
