+++ drivers/s390/char/sclp_mem.c
@@ -87,7 +87,7 @@ static int do_assign_storage(sclp_cmdw_t cmd, u16 rn)
 		break;
 	}
 out:
-	free_page((unsigned long)sccb);
+	free_page(sccb);
 	return rc;
 }
 
@@ -136,7 +136,7 @@ static int sclp_attach_storage(u8 id)
 		break;
 	}
 out:
-	free_page((unsigned long)sccb);
+	free_page(sccb);
 	return rc;
 }
 
@@ -393,7 +393,7 @@ static int __init sclp_detect_standby_memory(void)
 		goto out;
 	sclp_add_standby_memory();
 out:
-	free_page((unsigned long)sccb);
+	free_page(sccb);
 	return rc;
 }
 __initcall(sclp_detect_standby_memory);
