+++ drivers/s390/char/sclp_pci.c
@@ -66,7 +66,7 @@ static int do_pci_configure(sclp_cmdw_t cmd, u32 fid)
 		break;
 	}
 out:
-	free_page((unsigned long) sccb);
+	free_page(sccb);
 	return rc;
 }
 
@@ -174,7 +174,7 @@ int sclp_pci_report(struct zpci_report_error_header *report, u32 fh, u32 fid)
 	}
 
 out_free_req:
-	free_page((unsigned long) sccb);
+	free_page(sccb);
 out_unregister:
 	sclp_unregister(&sclp_pci_event);
 out_unlock:
