+++ drivers/s390/net/qeth_l2_main.c
@@ -599,7 +599,7 @@ static int qeth_l2_pnso(struct qeth_card *card, u8 oc, int cnc,
 	if (rc)
 		QETH_CARD_TEXT_(card, 2, "PNrp%04x", rr->response.code);
 
-	free_page((unsigned long)rr);
+	free_page(rr);
 	return rc;
 }
