+++ drivers/scsi/megaraid/megaraid_sas_base.c
@@ -8056,7 +8056,7 @@ skip_firing_dcmds:
 				if (is_vmalloc_addr(fusion->ld_drv_map[i]))
 					vfree(fusion->ld_drv_map[i]);
 				else
-					free_pages((ulong)fusion->ld_drv_map[i],
+					free_pages(fusion->ld_drv_map[i],
 						   fusion->drv_map_pages);
 			}
